+++ gulp/src/manageusers/Role.jsx
@@ -5,9 +5,9 @@ import {Link} from 'react-router';
 import Button from 'react-bootstrap/lib/Button';
 
 import {getCsrf} from 'util/cookie';
-import {reverseFormatActivityName} from 'util/reverseFormatActivityName';
-import {formatActivityNames} from 'util/formatActivityNames';
-import {buildCurrentActivitiesObject} from 'util/buildCurrentActivitiesObject';
+import {reverseFormatActivityName} from './reverseFormatActivityName';
+import {formatActivityNames} from './formatActivityNames';
+import {buildCurrentActivitiesObject} from './buildCurrentActivitiesObject';
 
 import HelpText from './HelpText';
 import ActivitiesAccordion from './ActivitiesAccordion';
