+++ gulp/src/manageusers/buildCurrentActivitiesObject.js
