+++ gulp/src/manageusers/formatActivityName.js
