+++ gulp/src/manageusers/formatActivityNames.js
