+++ gulp/src/manageusers/formatForMultiselect.js
