+++ gulp/src/manageusers/reverseFormatActivityName.js
