+++ gulp/src/manageusers/spec/formatActivityName.js
@@ -1,4 +1,4 @@
-import {formatActivityName} from '../../util/formatActivityName';
+import {formatActivityName} from '../formatActivityName';
 
 const input = 'read contact';
