+++ gulp/src/manageusers/spec/formatForMultiselectSpec.js
@@ -1,4 +1,4 @@
-import {formatForMultiselect} from '../../util/formatForMultiselect';
+import {formatForMultiselect} from '../formatForMultiselect';
 
 const availableUsersParsed = [
   {
